Some changes, failed

+++ Engine/Tables/Bishops.cs
@@ -2,7 +2,7 @@
 {
     public static class Bishops
     {
-        public static readonly int[] WhiteBishopTable = new int[64]
+        public static readonly double[] WhiteBishopTable = new double[64]
         {
             -20, -10, -10, -10, -10, -10, -10, -20,
             -10, 0, 0, 0, 0, 0, 0, -10,
@@ -14,7 +14,7 @@
             -20, -10, -10, -10, -10, -10, -10, -20
         };
 
-        public static readonly int[] BlackBishopTable = new int[64]
+        public static readonly double[] BlackBishopTable = new double[64]
         {
             -20, -10, -10, -10, -10, -10, -10, -20,
             -10, 5, 0, 0, 0, 0, 5, -10,
@@ -26,12 +26,12 @@
             -20, -10, -10, -10, -10, -10, -10, -20
         };
 
-        public static int GetWhiteSquareWeight(int square)
+        public static double GetWhiteSquareWeight(int square)
         {
             return WhiteBishopTable[square];
         }
 
-        public static int GetBlackSquareWeight(int square)
+        public static double GetBlackSquareWeight(int square)
         {
             return BlackBishopTable[square];
         }
